hardcode characteristic id in code & remove product_characteristic_int_value

+++ src/parse/db/entity/characteristic/category_characteristic.rs
@@ -5,12 +5,12 @@ use crate::schema::category_characteristic;
 pub struct CategoryCharacteristic {
     pub id: i32,
     pub category_id: i32,
-    pub characteristic_id: i32,
+    pub characteristic_id: i16,
 }
 
 #[derive(Insertable, Debug)]
 #[table_name = "category_characteristic"]
 pub struct NewCategoryCharacteristic {
     pub category_id: i32,
-    pub characteristic_id: i32,
+    pub characteristic_id: i16,
 }
