+++ src/parse/db/entity/characteristic/product_characteristic.rs
@@ -14,6 +14,6 @@ pub struct ProductCharacteristic {
 #[table_name = "product_characteristic"]
 pub struct NewProductCharacteristic {
     pub product_id: i32,
-    pub characteristic_id: i32,
+    pub characteristic_id: i16,
     pub value_id: i32,
 }
